Correct order of dropping certain database objects.
 * pyrseas/database.py (Database.diff_map): Drop functions before
   schemas and add dropping of types.
 * pyrseas/dbobject/column.py (ColumnDict.query): Use pg_get_expr() to
   get DEFAULT.
 * pyrseas/dbobject/constraint.py (ConstraintDict.diff_map): Skip
   domains when dropping CHECK constraints.
 * pyrseas/dbobject/dbtype.py (TypeDict.diff_map): Don'd drop types
   yet.  (TypeDict._drop): New method to actually drop the types.
 * pyrseas/dbobject/table.py (Sequence.get_dependent_table): Renamed
   from get_owner. Add query against pg_depend and pg_attrdef to
   determine the dependent table (when an owner table isn't present).
   (Sequence.to_map): Don't map dependent_table.
   (ClassDict._from_catalog): Change get_owner to get_dependent_table.
   (ClassDict.diff_map): Drop rules. If a sequence has a dependent
   table, postpone dropping it until after the tables are dropped.

+++ pyrseas/dbobject/constraint.py
@@ -409,7 +409,8 @@ class ConstraintDict(DbObjectDict):
                 elif turn == 2:
                     continue
                 # if missing, drop it
-                if (sch, tbl, cns) not in inconstrs:
+                if (sch, tbl, cns) not in inconstrs \
+                        and not hasattr(constr, 'target'):
                     stmts.append(constr.drop())
             # check input constraints
             for (sch, tbl, cns) in inconstrs.keys():
